Fixing final adjusts

+++ src/main/java/br/com/mybudget/userdashboard/model/dto/TotalBudgetDTO.java
@@ -19,6 +19,6 @@ import lombok.ToString;
 @ToString
 public class TotalBudgetDTO {
 	
-	private BigDecimal totalBudgetAmount;
-	private BigDecimal totalValueSaved;
+	private double totalBudgetAmount;
+	private double totalValueSaved;
 }
